+++ FuturaeKit.xcframework/ios-arm64/FuturaeKit.framework/Headers/FTRConfig.h
@@ -23,6 +23,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, copy, readonly) NSString *sdkId;
 @property (nonatomic, copy, readonly) NSString *baseUrl;
 @property (nonatomic, copy, readonly) NSString *locale;
+@property (nonatomic, copy, readonly) NSString *appGroup;
 @property (nonatomic, readonly) BOOL pinCerts;
 @property (nonatomic, strong, readonly) NSArray *capabilities;
 @property (nonatomic, strong, readonly) FTRKeychainConfig *keychain;
@@ -71,6 +72,58 @@ NS_ASSUME_NONNULL_BEGIN
                        pinCerts:(BOOL)pinCerts
                        keychain:(FTRKeychainConfig *)keychain
               lockConfiguration: (LockConfiguration * )lockConfiguration;
+
++ (instancetype)configWithSdkId:(NSString *)sdkId
+                         sdkKey:(NSString *)sdkKey
+              lockConfiguration: (LockConfiguration * )lockConfiguration
+                       appGroup:(NSString *)appGroup;
+
++ (instancetype)configWithSdkId:(NSString *)sdkId
+                         sdkKey:(NSString *)sdkKey
+                       keychain:(FTRKeychainConfig *)keychain
+              lockConfiguration: (LockConfiguration * )lockConfiguration
+                       appGroup:(NSString *)appGroup;
+
++ (instancetype)configWithSdkId:(NSString *)sdkId
+                         sdkKey:(NSString *)sdkKey
+                        baseUrl:(NSString *)baseUrl
+              lockConfiguration: (LockConfiguration * )lockConfiguration
+                       appGroup:(NSString *)appGroup;
+
++ (instancetype)configWithSdkId:(NSString *)sdkId
+                         sdkKey:(NSString *)sdkKey
+                        baseUrl:(NSString *)baseUrl
+                       keychain:(FTRKeychainConfig *)keychain
+              lockConfiguration: (LockConfiguration * )lockConfiguration
+                       appGroup:(NSString *)appGroup;
+
++ (instancetype)configWithSdkId:(NSString *)sdkId
+                         sdkKey:(NSString *)sdkKey
+                       pinCerts:(BOOL)pinCerts
+              lockConfiguration: (LockConfiguration * )lockConfiguration
+                       appGroup:(NSString *)appGroup;
+
++ (instancetype)configWithSdkId:(NSString *)sdkId
+                         sdkKey:(NSString *)sdkKey
+                       pinCerts:(BOOL)pinCerts
+                       keychain:(FTRKeychainConfig *)keychain
+              lockConfiguration: (LockConfiguration * )lockConfiguration
+                       appGroup:(NSString *)appGroup;
+
++ (instancetype)configWithSdkId:(NSString *)sdkId
+                         sdkKey:(NSString *)sdkKey
+                        baseUrl:(NSString *)baseUrl
+                       pinCerts:(BOOL)pinCerts
+              lockConfiguration: (LockConfiguration * )lockConfiguration
+                       appGroup:(NSString *)appGroup;
+
++ (instancetype)configWithSdkId:(NSString *)sdkId
+                         sdkKey:(NSString *)sdkKey
+                        baseUrl:(NSString *)baseUrl
+                       pinCerts:(BOOL)pinCerts
+                       keychain:(FTRKeychainConfig *)keychain
+              lockConfiguration: (LockConfiguration * )lockConfiguration
+                       appGroup:(NSString *)appGroup;
 
 - (BOOL)isValid;