+++ FuturaeKit.xcframework/ios-arm64/FuturaeKit.framework/Headers/FuturaeKit-Swift.h
@@ -260,6 +260,7 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 #if __has_warning("-Watimport-in-framework-header")
 #pragma clang diagnostic ignored "-Watimport-in-framework-header"
 #endif
+@import Foundation;
 @import ObjectiveC;
 #endif
 
@@ -312,6 +313,13 @@ SWIFT_CLASS_PROPERTY(@property (nonatomic, class, readonly, copy) NSString * _No
 SWIFT_CLASS("_TtC10FuturaeKit14DeviceSecurity")
 @interface DeviceSecurity : NSObject
 + (JailbreakStatus * _Nonnull)jailbreakStatus SWIFT_WARN_UNUSED_RESULT;
+- (nonnull instancetype)init OBJC_DESIGNATED_INITIALIZER;
+@end
+
+
+SWIFT_CLASS("_TtC10FuturaeKit13JWEDecryption")
+@interface JWEDecryption : NSObject
++ (NSArray<NSDictionary<NSString *, id> *> * _Nullable)decryptWithKey:(NSData * _Nonnull)key encryptedData:(NSString * _Nonnull)encryptedData SWIFT_WARN_UNUSED_RESULT;
 - (nonnull instancetype)init OBJC_DESIGNATED_INITIALIZER;
 @end